feat(cotizacion end, vehicles on)!

+++ api/src/controllers/quoter.controller.ts
@@ -162,7 +162,8 @@ export const updateQuoter = async (req: Request, res: Response) => {
             id_grua_gt,
             detalles,
             serie,
-            vin
+            vin,
+            aprobada = false
         } = req.body;
 
         const quoter = await Quoter.findOneBy({ id_cotizacion: Number(id) });
@@ -212,15 +213,21 @@ export const updateQuoter = async (req: Request, res: Response) => {
                 modelo: modelo ?? quoter.modelo,
                 anio: anio ?? quoter.anio,
                 serie: serie ?? quoter.serie,
-                vin: vin ?? quoter.vin
+                vin: vin ?? quoter.vin,
+                aprobada: aprobada ?? quoter.aprobada
             }
         );
 
-        await QuoterDetail.createQueryBuilder().delete().where({ quoter: quoter }).execute();
-        await insertDetail(quoter, detalles);
-        if ((update?.affected ?? 0) > 0) return res.json(quoter);
+        if (!aprobada) {
+            await QuoterDetail.createQueryBuilder().delete().where({ quoter: quoter }).execute();
+            await insertDetail(quoter, detalles);
 
-        return res.status(203).json({ message: 'No se pudo actualizar la cotización' });
+            if ((update?.affected ?? 0) > 0) return res.json(quoter);
+            return res.status(203).json({ message: 'No se pudo actualizar la cotización' });
+        } else {
+            if ((update?.affected ?? 0) > 0) return res.json({ message: 'Cotización aprobada' });
+            return res.status(203).json({ error: true, message: 'No se pudo aprobar la cotización' });
+        }
     } catch (error) {
         return res.status(500).json({ message: (error as Error).message });
     }
@@ -275,5 +282,56 @@ export const generatePdf = async (req: Request, res: Response) => {
         zip.finalize();
     } catch (error) {
         return res.status(500).json({ message: (error as Error).message, res: 'papas' });
+    }
+};
+
+export const getQuoterPaginatedData = async (req: Request, res: Response) => {
+    try {
+        const { pageSize = 100, current = 1, sortField = 'id_cotizacion', sortOrder = 'ASC', filter = '' } = req.body;
+
+        // Validar que los valores recibidos sean correctos
+        const validFields = ['id_cotizacion', 'fecha_creacion', 'vendedor', 'cliente', 'aprobada', 'id_cliente']; // Lista de campos válidos para ordenar
+        if (!validFields.includes(sortField)) return res.status(203).json({ message: 'Campo de orden inválido' });
+
+        const validDirections = ['ASC', 'DESC'];
+        if (!validDirections.includes(sortOrder.toUpperCase())) return res.status(203).json({ message: 'Dirección de orden inválida' });
+
+        // Crear la consulta base
+        const query = Quoter.createQueryBuilder('quoter')
+            .leftJoinAndSelect('quoter.cliente', 'cliente')
+            .leftJoinAndSelect('quoter.vendedor', 'vendedor');
+
+        // Aplicar filtro si es necesario
+        if (filter != '') {
+            if (isNaN(Number(filter))) {
+                query
+                    .where('quoter.fecha_creacion LIKE :filter', { filter: `%${filter}%` })
+                    .orWhere('vendedor.nombre LIKE :filter', { filter: `%${filter}%` })
+                    .orWhere('cliente.cliente LIKE :filter', { filter: `%${filter}%` });
+            } else {
+                query.where('cliente.id_cliente = :id', { id: Number(filter) });
+            }
+        }
+
+        if (sortField.includes('cliente')) query.orderBy(`cliente.${sortField}`, sortOrder.toUpperCase() as 'ASC' | 'DESC');
+        else if (sortField.includes('vendedor')) query.orderBy(`vendedor.nombre`, sortOrder.toUpperCase() as 'ASC' | 'DESC');
+        else query.orderBy(`quoter.${sortField}`, sortOrder.toUpperCase() as 'ASC' | 'DESC');
+
+        // Aplicar paginación
+        query.skip((current - 1) * pageSize).take(pageSize);
+
+        // Ejecutar la consulta
+        const [data, total] = await query.getManyAndCount();
+
+        // Retornar los datos paginados
+        return res.status(200).json({
+            data,
+            total,
+            current: Number(current),
+            pageSize: Number(pageSize),
+            totalPages: Math.ceil(total / pageSize)
+        });
+    } catch (error) {
+        return res.status(500).json({ message: (error as Error).message });
     }
 };
